NotificationService, Dao en JPA gemaakt + opzet observerpattern

+++ sdp2/src/main/java/repository/NotificationDao.java
@@ -4,5 +4,5 @@ import domein.Notification;
 import jakarta.persistence.EntityNotFoundException;
 
 public interface NotificationDao extends GenericDao<Notification> {
-    Notification getNotificationById(int id) throws EntityNotFoundException;
+    Notification getNotificationById(long id) throws EntityNotFoundException;
 }
